Switch to python-pptx for presentation management
After doing some more searching, I found [python-pptx][] which is doing
exactly what I want from the stand point of presentation management.
Therefore, we should just use that.  It can easily insert the images as
backgrounds and is has features for adding the slide notes which means I
won't have to much with the XML myself.

[python-pptx]: https://python-pptx.readthedocs.io/en/latest/index.html