Update noxfile to read from pyproject.toml
We need to pull the dependencies from the toml file instead of the
setup.cfg.  This silences a warning from setuptools.  We also refactor a
bit to combine the lint stages and use a better syntax to prevent the
main toml structure from getting mangled.  We also switch to reusing the
virtual environments by default.